% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetActivity.R
\docType{package}
\name{NetActivity}
\alias{NetActivity}
\title{NetActivity: compute gene set scores from a deep learning framework}
\description{
NetActivity enables to compute gene set scores from previously trained
sparsely-connected autoencoders. The package contains a function to prepare
the data (`prepareSummarizedExperiment`) and a function to compute the gene
set scores (`computeGeneSetScores`). The package `NetActivityData` contains
different pre-trained models to be directly applied to the data.
Alternatively, the users might use the package to compute gene set scores
using custom models.
}
